% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear.code.R
\name{clear.code}
\alias{clear.code}
\title{Creates a file copying current code without comments and tabuations}
\usage{
clear.code()
}
\value{
clear code to be executed
}
\description{
Creates a file copying current code without comments and tabuations
}
\examples{

\donttest{
# Copy and paste all the following code in RStudio
}
1+1

clear.code()

}
