% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copypaste.R
\name{copypaste}
\alias{copypaste}
\title{A function to import copied data in R via clipboard, the use of Addins is also highly recommended}
\usage{
copypaste()
}
\value{
A data frame object called "x" in .GlobalEnv
}
\description{
A function to import copied data in R via clipboard, the use of Addins is also highly recommended
}
\examples{

#Copy (Ctrl+C) a excel/google sheets/ online table such as:
#https://en.wikipedia.org/wiki/Standard_normal_table

#Run
#copypaste()

#Check your data
#x

}
