% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empty_df.R
\name{empty_df}
\alias{empty_df}
\title{Create an empty data.frame with specified number of columns and column names}
\usage{
empty_df(ncol, colnames)
}
\arguments{
\item{ncol}{integer; number of columns}

\item{colnames;}{character vector equal length to ncol; column names}
}
\value{
empty data.frame
}
\description{
Create an empty data.frame with specified number of columns and column names
}
