% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_file_list.R
\name{get_file_list}
\alias{get_file_list}
\title{Makes a tidy list of files within a directory.  Non-recursive, doesn't include subdirectories.}
\usage{
get_file_list(dir, pattern = NULL)
}
\arguments{
\item{dir}{character string; path to directory containing file paths to be stored}

\item{pattern}{character string; regular expression by which to match files within `dir`}
}
\value{
named list; values are full paths to all files within `dir` (or files which match `pattern`) and are named after their file.
}
\description{
Makes a tidy list of files within a directory.  Non-recursive, doesn't include subdirectories.
}
