% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r.args.R
\name{r.args}
\alias{r.args}
\title{Displays all arguments and its values from a selected R function}
\usage{
r.args()
}
\value{
arguments ready to run
}
\description{
Displays all arguments and its values from a selected R function
}
\examples{
\donttest{
# Copy and paste the function in RStudio, select the text and push Addins/Read Arguments
}

rnorm(n=100)

}
