% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libraries.R
\name{libraries}
\alias{libraries}
\title{Compiles and installs all libraries in the current RStudio script where the word library is written}
\usage{
libraries()
}
\value{
libraries compiled
}
\description{
Compiles and installs all libraries in the current RStudio script where the word library is written
}
\examples{
\donttest{
# Copy and paste all the following code in RStudio
}
loaded_pkgs<-(.packages())

\donttest{
# library(ggplot2)
# library(dplyr)
}
libraries()

loaded_pkgs2<-(.packages())


loaded_pkgs2[!loaded_pkgs2 \%in\% loaded_pkgs]


}
