% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libraries.R
\name{libraries}
\alias{libraries}
\title{compile all libraries in a ".R" file, even if is not installed}
\usage{
libraries()
}
\value{
libraries compiled
}
\description{
compile all libraries in a ".R" file, even if is not installed
}
\examples{

library(stats)

#libraries()

}
