% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r.args.R
\name{r.args}
\alias{r.args}
\title{Displays all arguments and its values from a selected R function}
\usage{
r.args()
}
\value{
arguments ready to run
}
\description{
Displays all arguments and its values from a selected R function
}
\examples{

#Copy and paste the following code in RStudio

#Select the following line of code & push Addins -> shortcuts / Read Arguments
rnorm(n=100)

#Result
n=100;mean=1;sd=0

}
